requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
click==6.6                # via pip-tools
coverage==4.2             # via pytest-cov
defusedxml==0.4.1         # via python3-openid
django-allauth==0.27.0
django-bootstrap3==7.1.0
django-filter==0.15.0
django==1.10.1
djangorestframework==3.4.7
factory-boy==2.7.0
fake-factory==0.7.2       # via factory-boy
first==2.0.1              # via pip-tools
markdown==2.6.6
oauthlib==2.0.0           # via requests-oauthlib
pip-tools==1.7.0
py==1.4.31                # via pytest
pytest-cov==2.3.1
pytest-django==3.0.0
pytest-splinter==1.7.6
pytest==3.0.2             # via pytest-cov, pytest-django, pytest-splinter
python-dateutil==2.5.3    # via fake-factory
python3-openid==3.0.10    # via django-allauth
requests-mock==1.1.0
requests-oauthlib==0.7.0  # via django-allauth
requests==2.11.1
selenium==2.53.6          # via pytest-splinter, splinter
six==1.10.0               # via fake-factory, pip-tools, python-dateutil, requests-mock
splinter==0.7.4           # via pytest-splinter

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools                # via pytest-splinter
